# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DiagnosticsResponse": {
        "additionalProperties": false,
        "properties": {
          "adapter": {
            "$ref": "#/components/schemas/ComponentReadiness"
          },
          "api_version": {
            "const": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ComponentReadiness"
          },
          "database": {
            "$ref": "#/components/schemas/ComponentReadiness"
          },
          "identity": {
            "$ref": "#/components/schemas/ComponentReadiness"
          },
          "recommendations": {
            "$ref": "#/components/schemas/ComponentReadiness"
          }
        },
        "required": [
          "api_version",
          "database",
          "data",
          "identity",
          "adapter",
          "recommendations"
        ],
        "title": "DiagnosticsResponse",
        "type": "object"
      },
...
      "DraftStateResponse": {
        "additionalProperties": false,
        "properties": {
          "accepted_picks": {
            "title": "Accepted Picks",
            "type": "integer"
          },
          "current_pick": {
            "title": "Current Pick",
            "type": "integer"
          },
          "dataset_version": {
            "title": "Dataset Version",
            "type": "string"
          },
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "feature_version": {
            "title": "Feature Version",
            "type": "string"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "title": "Issues",
            "type": "array"
          },
          "league_id": {
            "title": "League Id",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "reconciliation_state": {
            "title": "Reconciliation State",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "unresolved_observations": {
            "title": "Unresolved Observations",
            "type": "integer"
          }
        },
        "required": [
          "draft_id",
          "league_id",
          "status",
          "reconciliation_state",
          "revision",
          "current_pick",
          "dataset_version",
          "feature_version",
          "model_version",
          "accepted_picks",
          "unresolved_observations",
          "issues"
        ],
        "title": "DraftStateResponse",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "NFL Fantasy Assistant",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/diagnostics": {
      "get": {
        "operationId": "diagnostics_v1_diagnostics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostics"
      }
    },
    "/v1/drafts": {
      "post": {
        "operationId": "create_draft_v1_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Draft"
      }
    },
    "/v1/drafts/{draft_id}": {
      "get": {
        "operationId": "get_draft_v1_drafts__draft_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Draft"
      }
    },
    "/v1/drafts/{draft_id}/events": {
      "post": {
        "operationId": "ingest_event_v1_drafts__draft_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest Event"
      }
    },
    "/v1/drafts/{draft_id}/recommendations": {
      "get": {
        "operationId": "recommendations_v1_drafts__draft_id__recommendations_get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Recommendations"
      }
    },
    "/v1/drafts/{draft_id}/snapshot": {
      "post": {
        "operationId": "reconcile_snapshot_v1_drafts__draft_id__snapshot_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reconcile Snapshot"
      }
    },
...
    "/v1/leagues": {
      "post": {
        "operationId": "create_league_v1_leagues_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeagueCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create League"
      }
    }
  }
}